user can create and edit records

+++ models/record.py
@@ -8,6 +8,6 @@ class Record(BaseModel):
     date = pw.DateField(null=False)
     account = pw.ForeignKeyField(Account, backref="records")
     assessment = pw.ForeignKeyField(Assessment, backref="records", on_delete="CASCADE")
-    amount = pw.DecimalField(null=False, decimal_places=2)
+    amount = pw.IntegerField(null=False)
     description = pw.TextField(null=True)
     reference = pw.TextField(null=True)
